fix: juntando todos os usecases de cliente em módulo único

+++ src/usecase/cliente/deletar-cliente-usecase.ts
@@ -7,9 +7,18 @@ class DeletarClienteUseCase {
     this.clienteRepository = clienteRepository;
   }
 
-  async deletarCliente(id: string): Promise<void> {
-    await this.clienteRepository.removerCliente(id);
+  async deletarCliente(id: string): Promise<boolean | Error> {
+    const existe = await this.clienteRepository.buscarClientePorId(id);
+    if (!existe) {
+      return new Error('Cliente não encontrado');
+    }
+    try {
+      await this.clienteRepository.removerCliente(id);
+    } catch (error: any) {
+      return new Error(error.message);
+    }
+    return true;
   }
 }
 
-export {DeletarClienteUseCase};
+export { DeletarClienteUseCase };
